Use specifc gcn.alert_logger not general alertstreams.alert_logger
Each AlertStream subclass may use a specific kafka client with
it's own way of making the alert available to consumers. So, it
doesn't make sense to supply a general example alert_logger at
the alertstreams module level. Example loggers belong in the
AlertStream subclass, paired with it's specific kafka client
(in this case gcn-kafka).